Added AcquireTree() in Nlp.java
AcquireTree() makes use of CoreNLP's TreeAnnotation.class in order to 
show a parse tree in the format of (ROOT (S ( ... ) ) ).
Still contains the word ROOT, can be removed if requested.
The method also creates a tree.txt file in case you to read the
arraylist of sentences in strings in tree annotation.
Also made edits to ReaderWrite.java